requirements.txt
# Generated from environment.yml
# Conda dependencies converted to pip where possible
accelerate==1.7.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
bitsandbytes==0.45.5
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.13
ffmpy==0.6.0
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
gradio==4.16.0
gradio-client==0.8.1
h11==0.14.0
hf-xet==1.1.4
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.33.0
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
latex2mathml==3.78.0
llava==1.2.2.post1
markdown-it-py==3.0.0
markdown2==2.5.3
markupsafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.43.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
orjson==3.10.18
packaging==25.0
pandas==2.3.0
parso==0.8.4
peft==0.15.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==25.1.1
platformdirs==4.3.8
prompt-toolkit==3.0.51
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pygments==2.19.1
pydantic==2.11.7
pydantic-core==2.33.2
pydub==0.25.1
pyparsing==3.2.3
pysocks==1.7.1
python-dateutil==2.9.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
ruff==0.12.0
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
sentencepiece==0.1.99
setuptools==78.1.1
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
stack_data==0.6.3
starlette==0.46.2
svgwrite==1.4.3
sympy==1.13.1
threadpoolctl==3.6.0
timm==1.0.8
tokenizers==0.21.1
tomlkit==0.12.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
triton==2.1.0
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
wheel==0.45.1
zipp==3.23.0